Concerns #25 - Moved the ema utils into the core, as a moving-average module
